requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
requests
prometheus-flask-exporter
gunicorn
bandit
safety
semgrep
anthropic
openai
python-dotenv
pymongo
mongomock
